fix roi_info setting when use set_project_info in analysis
