+++ package/PartSegCore/analysis/calculate_pipeline.py
@@ -8,8 +8,8 @@ from PartSegCore.analysis.algorithm_description import analysis_algorithm_dict
 from PartSegCore.analysis.analysis_utils import SegmentationPipeline
 from PartSegCore.io_utils import HistoryElement
 from PartSegCore.mask_create import calculate_mask
+from PartSegCore.project_info import AdditionalLayerDescription
 from PartSegCore.segmentation import RestartableAlgorithm
-from PartSegCore.segmentation.algorithm_base import AdditionalLayerDescription
 from PartSegImage import Image
