+++ package/PartSegCore/project_info.py
@@ -1,15 +1,38 @@
 import sys
-from typing import Optional
+from dataclasses import dataclass
+from typing import Dict, Optional
 
 import numpy as np
 
 from PartSegCore.roi_info import ROIInfo
+from PartSegCore.utils import numpy_repr
 from PartSegImage import Image
 
 if sys.version_info.minor < 8:
     from typing_extensions import Protocol, runtime_checkable
 else:
     from typing import Protocol, runtime_checkable
+
+
+@dataclass
+class AdditionalLayerDescription:
+    """
+    Dataclass
+
+    :ivar numpy.ndarray data: layer data
+    :ivar str layer_type: napari layer type
+    :ivar str name: layer name
+    """
+
+    data: np.ndarray
+    layer_type: str
+    name: str = ""
+
+    def __repr__(self):
+        return (
+            f"AdditionalLayerDescription(data={numpy_repr(self.data)},"
+            f" layer_type='{self.layer_type}', name='{self.name}')"
+        )
 
 
 @runtime_checkable
@@ -29,6 +52,7 @@ class ProjectInfoBase(Protocol):
     image: Image
     roi: np.ndarray
     roi_info: ROIInfo = ROIInfo(None)
+    additional_layers: Dict[str, AdditionalLayerDescription] = dict()
     mask: Optional[np.ndarray]
     errors: str = ""
     points: Optional[np.ndarray] = None
